+++ CHANGELOG.rst
@@ -14,6 +14,10 @@ and this project adheres to `Semantic Versioning`_.
 Unreleased_
 -----------
 
+Changed
+.......
+* `withSample` now takes a probability argument, and include the samples in the generation
+
 Added
 .....
 * ``andThen`` operator to make easier building complex generation pipeline